[Refactor] configuration 폴더 명 수정
